Fix broken SRW_to_wind (#60)
* Update ResourceTools.py

When fetching files from the NSRDB using `PySAM.ResourceTools.FetchResourceFiles(tech='pv')`, the returned csv file lists the time zone = 0, or in UTC time. According to the [NSRDB API documentation](https://nsrdb.nrel.gov/data-sets/api-instructions.html), it states that the `utc` parameter in the API call must be set to false:
```
# Specify Coordinated Universal Time (UTC), 'true' will use UTC, 'false' will use the local time zone of the data.
  # NOTE: In order to use the NSRDB data in SAM, you must specify UTC as 'false'. SAM requires the data to be in the
  # local time zone.
```
However, in the `_NSRDB_worker` function in `ResourceTools` the `lookup_query_url` variable does not include the `&utc=false` parameter

* Fix broken SRW_to_wind and wind local TZ

SRW_to_wind_data was not working for two reasons: capitalization of the field names, and python is unable to convert blank strings to float

Also, the _wtk_worker was fetching data in UTC time and not in local time, so I fixed the API query URL to fetch local time. This is now giving results consistent with the SAM GUI.

* Update ResourceTools.py SRW_to_wind_data() to support capitalized field names

Sample wind files that come with SAM use capitalized column labels Temperature, Pressure, etc.

Co-authored-by: Paul Gilman <cpaulgilman@gmail.com>